remove completeness crap
