+++ src/Pim/Bundle/EnrichBundle/Resources/config/normalizers.yml
@@ -52,6 +52,13 @@ services:
             - '@pim_enrich.provider.form.chained'
             - '@pim_catalog.localization.localizer.converter'
             - '@pim_enrich.converter.standard_to_enrich.product_value'
+            - '@pim_catalog.object_manager.product'
+            - '@pim_catalog.manager.completeness'
+            - '@pim_catalog.repository.channel'
+            - '@pim_catalog.filter.chained'
+            - '@pim_enrich.normalizer.completeness_collection'
+            - '%pim_catalog_product_storage_driver%'
+            - '@pim_user.context.user'
         tags:
             - { name: pim_internal_api_serializer.normalizer }
 
@@ -63,7 +70,7 @@ services:
     pim_enrich.normalizer.completeness_collection:
         class: '%pim_enrich.normalizer.completeness_collection.class%'
         arguments:
-            - '@pim_internal_api_serializer'
+            - '@pim_enrich.normalizer.completeness'
 
     pim_enrich.normalizer.version:
         class: '%pim_enrich.normalizer.version.class%'
